Move Txid into types
